update and insert tests working
