--- pkg/plugins/pg/destination/destination_v1_integration_test.go
@@ -1,1 +1,0 @@
-package destination
